Try to improve the solver's default solution
Set the solver flag SOLVER_FLAG_FOCUS_BEST.  This helps in the
following situation: A required package has an update available.  The
user chooses uninstall.  Without the flag the default solution is to
keep the (outdated) package.  With the flag the default is to upgrade.
It remains to be seen whether the flag has any undesired effects.