Remove PrereqChecker::upgrade
Also remove the 'update' parameter from SolverSolution::update.

These were introduced so that the SOLVER_UPDATE | SOLVER_SOLVABLE_ALL
command could be sent to the solver if and only if 'Keep' was not
globally selected.  But that command doesn't seem to be needed, so we
just remove it and simplify the code.